Refine documentation policy
The current documentation policy is somewhat abstract. Here we try to
make it more concrete while retaining the flexibilty of the old one in
a mostly backwards-compatible way. Most importantly, we specify how to
document traits, enums, structs and functions. The hope is also to
simplify the documentation process, especially for new and occasional
contributers.